+++ inst/resources/init.R
@@ -181,10 +181,10 @@ local({
     ## an 'installed from source' version
 
     ## -- InstallAgent -- ##
-    installAgent <- "InstallAgent: packrat 0.4.8-25"
+    installAgent <- 'InstallAgent: packrat 0.4.8-26'
 
     ## -- InstallSource -- ##
-    installSource <- "InstallSource: source"
+    installSource <- 'InstallSource: source'
 
     packratDescPath <- file.path(lib, "packrat", "DESCRIPTION")
     DESCRIPTION <- readLines(packratDescPath)